YoYoGames lies and GameMaker is broken
I had a nice menu setup that relied on object inheritance, but inheritance is really weird in GML.  I switched the instances to structs and it was even worse.  Everything that should have worked did not work.  Thank you GMS2, I will never use you again after this.

All that is left from what I worked on is the draw_text_tagged script.